+++ spec/lib/oauth/helpers/uri_checker_spec.rb
@@ -78,6 +78,12 @@ module Doorkeeper::OAuth::Helpers
         expect(described_class).not_to be_matches(uri, client_uri)
       end
 
+      it "allows state query parameter" do
+        uri = "http://app.co/?state=hello"
+        client_uri = "http://app.co"
+        expect(described_class).to be_matches(uri, client_uri)
+      end
+
       it "doesn't allow non-matching domains through" do
         uri = "http://app.abc/?query=hello"
         client_uri = "http://app.co"